refactor: create base transaction class
